social-links-library/scripts/requirements.txt
requests
beautifulsoup4
PyGithub
python-dotenv 